Added hardware documentation (#209)
* Added hardware documentation

* replaced webp with svg

* Update schem.svg

* Updated KiCad scheme

Based on the professor’s message on Teams, I initially assumed that resizing the SVG was unnecessary, as no image had been embedded in the drawing. However, after several commits and failed pipeline runs, I realized that the SVG was indeed too large. I have now adjusted it and hope the issue is resolved.

* fixed broken link